LayoutTests/fast/js/dfg-inline-new-array-buffer-expected.txt

This tests that inlining correctly handles constant buffers.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS bar()[0] is 1
PASS bar()[1] is 2
PASS bar()[2] is 3
PASS bar()[3] is 4
PASS bar()[0] is 1
PASS bar()[1] is 2
PASS bar()[2] is 3
PASS bar()[3] is 4
PASS bar()[0] is 1
PASS bar()[1] is 2
PASS bar()[2] is 3
PASS bar()[3] is 4
PASS bar()[0] is 1
PASS bar()[1] is 2
PASS bar()[2] is 3
PASS bar()[3] is 4
PASS bar()[0] is 1
PASS bar()[1] is 2
PASS bar()[2] is 3
PASS bar()[3] is 4
PASS bar()[0] is 1
PASS bar()[1] is 2
PASS bar()[2] is 3
PASS bar()[3] is 4
PASS bar()[0] is 1
PASS bar()[1] is 2
PASS bar()[2] is 3
PASS bar()[3] is 4
PASS bar()[0] is 1
PASS bar()[1] is 2
PASS bar()[2] is 3
PASS bar()[3] is 4
PASS bar()[0] is 1
PASS bar()[1] is 2
PASS bar()[2] is 3
PASS bar()[3] is 4
PASS bar()[0] is 1
PASS bar()[1] is 2
PASS bar()[2] is 3
PASS bar()[3] is 4
PASS successfullyParsed is true

TEST COMPLETE
